Added correction of error line number for batch start line:
The `ERROR_LINE()` function returns a line number relative to the start
of the batch -- in the case of `sp_executesql`, the string passed to
that function.  This results in a complicated error line report, as the
problem batch must first be determined before the error line number
can come into play.  This change adds a variable that tracks the
starting of the batch (relative to a fixed point) and adjusts the
reported line of the error accordingly.

Best offered support by SQL Server is "line within batch" --
`ERROR_LINE()` only returns the number of the line (starting with 1)
within the batch.  This is true even at the top level (i.e. outside all
calls to `sp_executesql`).  The best available solution (without
expanding the scope of the `in_ddl_transaction` method) is to provide an
"error reference line" and give offsets relative to it.  An alternative
would be to accept a prefix argument and explicitly handle the line
counting to get exact line numbers.