Add framework for automating declaration-change migration
This commit adds a framework for automating the generation of the SQL
needed to implement changes to declarative files.  Classes can `include
XMigra::ImpdeclMigrationAdder::SupportedDatabaseObject` and declare one
or more supported YAML tags with `for_declarative_tagged`.  An instance
of the class (or two in the case of a modification) will be constructed
and asked to suggest SQL to accomplish the change.