Fix the "branchid" subcommand for git repositories
Branch ID is less meaningful for Git repositories than for Subversion
ones, but being able to identify the branch similarly to how it is
listed in the generated script is still somewhat useful.

`xmigra branchid` was failing because it was depending on knowing
whether the generating object was for `production` (meaningful on
`SchemaUpdater` but not on the base-class `SchemaManipulator`) which is
not a known value for the `branchid` subcommand.  Once that was
resolved, it also failed because it did not detect the Git-special
attribute value of "unspecified".