Fix read from child-result pipe to prevent deadlock
If the child process filled the message pipe, it would deadlock as the
parent/runner process waited for the child to complete and the child
waited for the parent to read from the pipe.  Now, the parent reads from
the pipe until EOF (which occurs if the child exits), and then waits on
the child process.