Add logging of an exception (with backtrace) to a file
This will write the exception message, type, and backtrace to the file
specified by the `XMIGRA_LOG_FILE` environment variable, if present.
This assists with debugging when it is unclear why an exception
occurred.  `XMigra.log_error` should be called whenever a "big tarp"
exception handler is implemented -- basically it is helpful whenever the
code would swallow an exception without reraising.