Correct "migration not changed" test for Git VCS
Because Git doesn't bind commits to branches (the concept doesn't even
make sense), there is no way to check if a migration file is "unchanged
since committed to the production branch" as can be done in Subversion.
The closest-to-equivalent check is that the migration file as it exists
in the working copy has the same content as in the head of the
centrallized master branch (specified with the `xmigra-master` Git
attribute). This leaves it up to hooks in the central repository or
the merge review process to ensure the migration files in the master
branch do not change.

Issues: #31